Make cleanup optional after an AIO build
If the first argument passed to build.sh is "true" cleanup the
python virtual environment at the end of the script.

For local development it is useful to keep the python virtual
environment after the workflow completes.
When locally testing build scripts it is help to cleanup the
virtual environment. Since this is less common, default to not
cleaning the environment.

+++ aio/build.sh
@@ -157,7 +157,10 @@ makensis grampsaio64.nsi
 # result is in mingw64/src
 
 # deactivate and delete the python virtual environment
-deactivate
-rm -rf $pythonvenv
+if [ "$1" = "true" ]; then
+    echo "post build cleanup"
+    deactivate
+    rm -rf $pythonvenv
+fi
 
 exit 0
